change int -> decimal for rating movie

+++ Infrastructure/Ef/DbEntities/DbRatingMovie.cs
@@ -7,7 +7,7 @@ namespace Infrastructure.Ef.DbEntities;
 public class DbRatingMovie
 {
 
-    public int Average_rating { get; set; }
+    public decimal Average_rating { get; set; }
     public int NumVote { get; set; }
     
     [Key]
